// src/app/shared/models/trading-account.model.ts
export class TradingAccount {
  account_id: number;
  login: string;
  at_id: number;
  eq_group_id: number;
  is_demo: number;
  account_type: string;
  currency_id: number;
  currency: string;
  currency_sign: string;
  account_type_title: string;
  account_type_name: string;
  balance: string;
  balance_formatted: string;
  credit: string;
  credit_formatted: string;
  equity: string;
  equity_formatted: string;
  map_id: string;
  password: string;
  investor_password: string;
}